+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Nothing yet
+### Changed
+- Added `done()` calls to support llama-rlsr 0.0.4
 
 ## [0.0.3] 2017-08-12
 ### Fixed